+++ app/api/js-api/lib/request.js
@@ -1,4 +1,6 @@
 // @flow
+import _ from 'lodash';
+import blakejs from 'blakejs';
 import https from 'https';
 import querystring from 'querystring';
 
@@ -14,20 +16,43 @@ export type RequestOptions = {
   },
 };
 
+const bytesToB16 = (bytes) => Buffer.from(bytes).toString('hex');
+const blake2b = (data) => blakejs.blake2b(data, null, 32);
+const encryptPasspharse = (passphrase) => bytesToB16(blake2b(passphrase));
+
 export const request = (httpOptions: RequestOptions, queryParams?: {}) => {
   return new Promise((resolve, reject) => {
     // Prepare request with http options and (optional) query params
     const options: RequestOptions = Object.assign({}, httpOptions);
+    let hasRequestBody = false;
     let requestBody = '';
     if (queryParams) {
-      requestBody = querystring.stringify(queryParams);
-      options.headers = {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'Content-Length': requestBody.length,
-      };
+      // Handle passphrase
+      if (_.has(queryParams, 'passphrase')) {
+        const passpharse = queryParams.passphrase;
+
+        // If passphrase is present it must be encrypted and included in options.path
+        if (passpharse !== null) {
+          const encryptedPassphrase = encryptPasspharse(passpharse);
+          options.path += `?passphrase=${encryptedPassphrase}`;
+        }
+
+        // Passphrase is never sent as a part of requestBody so we need to omit it
+        queryParams = _.omit(queryParams, 'passphrase');
+      }
+
+      // Handle other query params
+      if (_.size(queryParams) > 0) {
+        hasRequestBody = true;
+        requestBody = querystring.stringify(queryParams);
+        options.headers = {
+          'Content-Type': 'application/x-www-form-urlencoded',
+          'Content-Length': requestBody.length,
+        };
+      }
     }
     const request = https.request(options);
-    if (queryParams) { request.write(requestBody); }
+    if (hasRequestBody) { request.write(requestBody); }
     request.on('response', (response) => {
       let body = '';
       // Cardano-sl returns chunked requests, so we need to concat them