+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Changelog
 - Fixed failing acceptance-tests ([PR 454](https://github.com/input-output-hk/daedalus/pull/454))
 - Prevent application window auto focusing while running acceptance-tests ([PR 458](https://github.com/input-output-hk/daedalus/pull/458))
 - Limit and validate wallet's name maximum length ([PR 465](https://github.com/input-output-hk/daedalus/pull/465))
+- Fixed broken js-api passphrase handling ([PR 474](https://github.com/input-output-hk/daedalus/pull/474))
 
 ### Chores